Computer Vs Computer feature
The computer vs computer feature was implemented such that 2 computers of different or same difficulty/mode play against each other. Since both players are computers, the game occurs really fast, making it unable for us to see except we interrupt it and then we see the victories, defeats, and ties of both computers.